refactor: migrate labels-editor to bundled TypeScript controller
Port public/modules/ui/labels-editor.js to src/controllers/labels-editor.ts
following the migration guide (strict d3 v7: select/drag/pointer/line/
curveNatural named imports, event-arg drag and control-point handlers,
module-local line generator). The static #labelEditor toggle sections stay
in index.html; listeners wire once behind an initialized flag. Register as
Controllers.LabelsEditor and replace editLabel() at the editors.js call site.
