+++ src/controllers/index.ts
@@ -16,6 +16,7 @@ export const Controllers = createRegistry({
   HeightmapSelection: () => import("@/controllers/heightmap-selection").then(m => m.HeightmapSelection),
   HierarchyTree: () => import("@/controllers/hierarchy-tree").then(m => m.HierarchyTree),
   IceEditor: () => import("@/controllers/ice-editor").then(m => m.IceEditor),
+  LabelsEditor: () => import("@/controllers/labels-editor").then(m => m.LabelsEditor),
   LakesEditor: () => import("@/controllers/lakes-editor").then(m => m.LakesEditor),
   MarkersEditor: () => import("@/controllers/markers-editor").then(m => m.MarkersEditor),
   MarketDealsOverview: () => import("@/controllers/market-deals-overview").then(m => m.MarketDealsOverview),
